Got messaging working
In socket_service.dart, changed the class variables amd methods to be static, so that it can be accessed from the network state class

in main.dart, added the ability to pass the address text controller, so that the server can now be connected to a user defined ip:port. Also changed the message formating code. Going to work on the tcp python server and emulate exchanging data and text